pyproject.toml
[build-system]
requires = ["setuptools==80.9.0", "wheel==0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "cogamer"
version = "0.1.0"
description = "Autonomous Claude Code agents for CoGames"
requires-python = ">=3.12,<3.13"
dependencies = [
  "click>=8.1",
  "httpx>=0.27",
  "fastapi>=0.115",
  "fastapi-mcp>=0.1",
  "uvicorn>=0.34",
  "pydantic>=2.0",
  "boto3>=1.35",
  "rich>=13.0",
  "pyyaml>=6.0.3",
  "pyjwt[crypto]>=2.0",
  "mangum>=0.19",
  "cogames",
  "mettagrid",
]

[project.scripts]
cogamer = "cogamer.cli:main"

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-asyncio>=0.24", "pytest-timeout>=2.3", "ruff"]
infra = ["aws-cdk-lib>=2.170", "constructs>=10.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
timeout = 60
pythonpath = ["src"]
addopts = "--import-mode=importlib -m 'not e2e'"
markers = ["e2e: end-to-end tests requiring external services"]

[tool.ruff]
extend = "../../.ruff.toml"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["PLC0415", "E402"]
"src/cogamer/**/__init__.py" = ["F401"]
# Lazy imports for circular-dep avoidance, optional deps, and CLI structure
"src/cogamer/*.py" = ["PLC0415"]
"src/cogamer/**/*.py" = ["PLC0415"]
"src/cogamer/channel/mcp_server.py" = ["E402"]

[tool.uv.sources]
cogames = { workspace = true }
mettagrid = { workspace = true }
